requirements.txt
Flask
flask-cors
Pillow
opencv-python
matplotlib
numpy
scipy
scikit-learn
